refactor(workflow): modularize the tests folder
